666
Added yet another demonic concept to the batch, one thought to be long forgotten. Still in preliminary phase, with most of the cards in need of further probation, but with enough bulk to them as to serve as a functional deck.

Enter Siphantom, the one that was meant to be.