SYNCHRO SLOP
I forgot to give this deck good tuners, well, that's all done and over with. Blazon has been fully realized now, that makes 3 decks out of 7 now. Stelloy should be up next, that one still needs some kinks and bolts locked in. Fulmiknight, Siphant and Desceptim are still up for further revisions, but if nothing else, their core design seems to be quite up to speed, just need to give them the touch.
And by that I mean not make them so egregiously pasty and synthetic, though in the case of at least two of these, it cannot be helped.
You are not immune to Union/Gemini propaganda.